% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_boot}
\alias{plot_boot}
\title{Bootstrap distribution of the BER}
\usage{
plot_boot(x, bins = 40)
}
\arguments{
\item{x}{An object of class \code{baf_boot} or \code{bafratio}.}

\item{bins}{Number of histogram bins.}
}
\value{
A \code{ggplot} object.
}
\description{
Bootstrap distribution of the BER
}
\examples{
boot <- baf_bootstrap(sim_est$logRr, sim_est$seLogRr,
                      sim_nc$logRr, sim_nc$seLogRr, nBoot = 500, seed = 42)
plot_boot(boot)
}
