% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_qq}
\alias{plot_qq}
\title{Q-Q plot of standardized negative control residuals}
\usage{
plot_qq(x)
}
\arguments{
\item{x}{An object of class \code{baf_diag} or \code{bafratio}.}
}
\value{
A \code{ggplot} object.
}
\description{
Q-Q plot of standardized negative control residuals
}
\examples{
d <- baf_diagnostics(sim_nc$logRr, sim_nc$seLogRr, ncNames = sim_nc$outcome)
plot_qq(d)
}
