% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_loo}
\alias{plot_loo}
\title{Forest plot of leave-one-out BER sensitivity}
\usage{
plot_loo(x)
}
\arguments{
\item{x}{An object of class \code{baf_loo}, or a \code{bafratio} object created
with \code{loo = TRUE}.}
}
\value{
A \code{ggplot} object.
}
\description{
Forest plot of leave-one-out BER sensitivity
}
\examples{
loo <- baf_loo(sim_est$logRr, sim_est$seLogRr,
               sim_nc$logRr, sim_nc$seLogRr,
               ncNames = sim_nc$outcome, nBoot = 100, seed = 1)
plot_loo(loo)
}
